build: working t2rn
